level update

+++ CollisionRect.h
@@ -5,12 +5,12 @@
 class CollisionRect
 {
 public:
-	int x;
-	int y;
-	int w;
-	int h;
+	double x;
+	double y;
+	double w;
+	double h;
 
-	CollisionRect(int x, int y, int w, int h) :x(x), y(y), w(w), h(h) {}
+	CollisionRect(double x, double y, double w, double h) :x(x), y(y), w(w), h(h) {}
 
 	std::string toString() { return "Rect: x: " + std::to_string(x) + " y: " + std::to_string(y) + " w: " + std::to_string(w) + " h: " + std::to_string(h); }
 };
